+++ NEWS.md
@@ -1,5 +1,10 @@
 # Package: c3co.data
 
+## Version: 0.0.2-9000 [2018-01-31]
+
+* data("PSCBSdata") now got sample names in names() as well as in the
+  individual PairedPSCBS objects.
+
 ## Version: 0.0.2 [2017-05-15]
 
 * Move all datasets to 'data' directory